+++ backend/src/main/java/com/pag/backend/repository/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pag.backend.domain.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 
-	public List<Order> findAllByCustomerId(Long id);
+	public List<Order> findAllByCustomerId(Integer id);
 }
